+++ lib/src/email.dart
@@ -2,7 +2,7 @@
 ///
 /// Blank field - null value - is a valid input!
 ///
-/// If the field is mandatory, use [ReqEmail].
+/// If the field is mandatory, see [Req].
 class Email {
   /// Validates an optional email field using a regular expression suitable for
   /// manually entered emails.
@@ -14,29 +14,17 @@ class Email {
   /// more than 4 subdomains.
   ///
   /// **Note:** null value is a **valid input**, whereas the empty string _''_
-  /// **is not**. If the email field is mandatory, use [ReqEmail].
+  /// is not. If the email field is mandatory, see [Req].
   ///
   /// [malformed] is the error message in case of a malformed email.
-  Email({required String malformed})
-      : this.regex(
-          malformed: malformed,
-          matcher: _default,
-        );
-
-  /// Email validation with custom regex.
-  ///
-  /// [invalid] is the error message in case of a malformed email; [matcher], is
-  /// the custom regex for email validation.
-  Email.regex({required String malformed, required RegExp matcher})
-      : _malformed = malformed,
-        _matcher = matcher;
+  const Email({required String malformed}) : _malformed = malformed;
 
   final String _malformed;
-  final RegExp _matcher;
   // lazy loading (on-demand) initialization.
-  static late final RegExp _default = _initDefaultRegex();
+  static late final RegExp _matcher = _emailPattern();
 
-  static RegExp _initDefaultRegex() {
+  /// suitable pattern for a manually entered email.
+  static RegExp _emailPattern() {
     return RegExp(
       r'^[A-Z0-9._%+-]{1,64}@(?:[A-Z0-9-]{1,63}\.){1,8}[A-Z]{2,63}$',
       caseSensitive: false,
@@ -44,7 +32,7 @@ class Email {
     );
   }
 
-  /// Valid - returns null - if [email] is either valid or null.
+  /// Valid - returns null - if [email] is either well-formed or null.
   String? call(String? email) =>
       (email == null || _matcher.hasMatch(email)) ? null : _malformed;
 }