+++ lib/src/req.dart
@@ -2,9 +2,11 @@
 ///
 /// **Note:** both the **null** value and the empty string **''** are **invalid inputs**.
 class Req {
-  /// Validation for non-blank fields.
+  /// Non-blank field with an optional extra validation step.
   ///
-  /// [blank] is the error message in case of a null or empty value.
+  /// [val] is the validation step to be performed after ensuring that the field
+  /// was not left blank; [blank] is the error message in case of a null or
+  /// empty value.
   const Req({required String blank}) : _blank = blank;
 
   final String _blank;