+++ lib/src/trim.dart
@@ -2,16 +2,16 @@ import 'type.dart';
 
 /// Trimmed value.
 ///
-/// It does not validate anything! It just trims the incoming value and forwards
-/// it to the encapsulated validator.
+/// It does not validate anything! It just trims the inpu value before
+/// forwarding it.
 class Trim {
-  /// The value to be trimmed - without any leading and trailing whitspace.
-  Trim(this._validator);
+  /// The value to be trimmed - remove any leading and trailing whitspace.
+  Trim(this._val);
 
-  final Callor _validator;
+  final Callor _val;
 
   /// Trims the value and forwards it.
   ///
   /// In other words, removes any leading and trailing whitespace.
-  String? call(String? value) => _validator(value?.trim());
+  String? call(String? value) => _val(value?.trim());
 }